refactor: 약결합 구조 적용
