+++ src/main/java/ladder/domain/LineGenerator.java
@@ -1,0 +1,8 @@
+package ladder.domain;
+
+import java.util.List;
+
+public interface LineGenerator {
+
+    List<Boolean> generate(int numberOfPlayers);
+}
